refactor(backend): store task node timestamps as integer milliseconds
## Problem
RuntimeInfo.start_time/end_time and NodeActionEvent.ts were stored as
float Unix seconds (time.time()), surfacing in the dashboard JSON with a
decimal point (e.g. 1786900492.534511). Consumers expected integer
millisecond epoch timestamps (13-digit int), the conventional web/JS form.

## Solution
Switched the task time fields to int milliseconds (int(time.time()*1000)):
- domain/models.py: start_time/end_time: int | None; NodeActionEvent.ts: int.
- adapters/http/task/schemas.py: matching DTO types (int).
- task_graph_service.py: all write sites (start_time/end_time/ts/
  bbs_claim_at/attach_bbs_node RuntimeInfo) now int(time.time()*1000).
- tests/test_bbs_claim.py: assertion baseline uses int ms.

Harness SLA math is unaffected — it uses its own monotonic clock
(_dispatched_at), never reads node.start_time; so the ms change does not
alter timeout comparisons. Spec three-pack (plan/spec/tasks) synced: field
types int ms; clarified the harness-commented "start_time + sla" to reflect
the actual monotonic-clock implementation.

## Validation
- ruff clean on changed task module files.
- task_center + e2e + task_graph + test_bbs_claim: 101 passed (incl.
  test_claim_writes_owner_and_claim_at which compares bbs_claim_at; and
  graph-service tests asserting start_time/end_time non-null/round-trip).

+++ src/backend/src/agentclaw/community/adapters/http/task/schemas.py
@@ -117,7 +117,7 @@ class AcceptanceResultDTO(BaseModel):
 class NodeActionEventDTO(BaseModel):
     """节点动作级历史快照(诊断用;默认不序列化,include_action_log=true 时返回)。"""
     seq: int
-    ts: float
+    ts: int
     action: str
     loop_round: int = 0
     attempt: int = 0
@@ -129,8 +129,8 @@ class NodeActionEventDTO(BaseModel):
 class RuntimeInfoDTO(BaseModel):
     run_mode: str | None = None
     assignee: str | None = None
-    start_time: float | None = None
-    end_time: float | None = None
+    start_time: int | None = None
+    end_time: int | None = None
     output: dict[str, Any] = Field(default_factory=dict)
     acceptance_result: AcceptanceResultDTO | None = None
     extend_props: dict[str, Any] = Field(default_factory=dict)
